+++ app/javascript/packages/i18n/index.ts
@@ -29,8 +29,16 @@ const getPluralizationKey = (count: number): keyof PluralizedEntry =>
  *
  * @return Entry string or object.
  */
-const getEntry = (strings: Entries, key: string): Entry =>
-  Object.hasOwn(strings, key) ? strings[key] : key;
+function getEntry(strings: Entries, key: string): Entry {
+  if (Object.hasOwn(strings, key)) {
+    return strings[key];
+  }
+
+  // eslint-disable-next-line no-console
+  console.error(`Missing translation for key \`${key}\`.`);
+
+  return key;
+}
 
 /**
  * Returns true if the given entry is a pluralization entry, or false otherwise.