+++ app/javascript/packages/i18n/CHANGELOG.md
@@ -1,4 +1,6 @@
-# `Change Log`
+## Unreleased
+
+- Output console message when translation data missing for requested key.
 
 ## 1.0.1